requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
fastapi==0.113.0
python-dotenv
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
httplib2==0.22.0
httpx
idna==3.10
mutagen==1.47.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings
pydantic_core==2.33.1
pyparsing==3.2.3
requests==2.32.3
rsa==4.9
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn[standard]==0.30.3
python-multipart
discogs-client
aiofiles
musicbrainzngs
# pyjamendo # Optional Jamendo client

# Week 1 Dependencies - Database, Auth, and Core
sqlalchemy==2.0.35
alembic==1.14.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
redis==5.2.1
celery==5.4.0

# Story Protocol SDK dependencies
web3==7.6.1
eth-account==0.13.7
eth-utils==5.3.0

# Audio fingerprinting
pyacoustid==1.3.0

# Testing and Development
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1
mypy==1.14.1

# Additional utilities
python-dateutil==2.9.0
boto3==1.36.10  # For S3 file storage
sentry-sdk==2.21.0  # For error tracking
structlog==25.1.0  # For structured logging
